Fixed bug to prevent users from entering invalid directions.
The getDir rutine was directly storing in memory the value input by the user if it was valid, which made it possible for the users to change the snake-direction two times in a single tick and therefore making normally not-valid snake-directions valid. Now the function doesnt even stores the new direction, as that is done in the main, it just returns the user's input if its valid.